% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patchmax_r6_func.R
\name{build_graph_func}
\alias{build_graph_func}
\title{Build graph object used to calculate patch}
\usage{
build_graph_func(net, objective_field, sdw = 0, epw = 0)
}
\arguments{
\item{net}{igraph object}

\item{sdw}{numeric >= 0 controlling flexibility; 0 negates any adjustment
related to the objective.}

\item{epw}{numeric > 0 distance multiplier for traversing excluded stands}

\item{obj}{name of variable containing objective}
}
\value{
cpp graph object
}
\description{
Build graph object used to calculate patch
}
\details{
\code{epw} values less than 1 will preferentially select excluded stands
while values greater than 1 avoids excluded stands.
}
