% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patchmax_r6_func.R
\name{build_patch_func}
\alias{build_patch_func}
\title{Build patch of specified size while minimizing modified distance costs}
\usage{
build_patch_func(
  start_node,
  cpp_graph,
  net,
  a_max,
  a_min = -Inf,
  c_max = Inf,
  c_min = -Inf,
  c_enforce = TRUE
)
}
\arguments{
\item{cpp_graph}{graph object}

\item{a_max}{size of patch}

\item{v}{node id to start building patch}
}
\value{
data frame of nearest nodes arranged by distance
}
\description{
Build patch of specified size while minimizing modified distance costs
}
