% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patchmax_r6_func.R
\name{calc_adj_network_func}
\alias{calc_adj_network_func}
\title{Estimates stand adjacency and calculates edgewise distances}
\usage{
calc_adj_network_func(shp, St_id, calc_dist = FALSE)
}
\arguments{
\item{shp}{sf-type geometry with attributes}

\item{St_id}{vector of sf feature ideas (i.e., the stand id)}

\item{calc_dist}{logical for calculating edgewise distance in adjacency network}

\item{buf_dist}{numeric distance used to buffer geometry and calculate adjacency}
}
\value{
adjacency network saved as an igraph network object
}
\description{
Estimates stand adjacency and calculates edgewise distances
}
