% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patchmax_r6_func.R
\name{calc_patch_stats}
\alias{calc_patch_stats}
\title{Helper for reporting patch statistics in build function}
\usage{
calc_patch_stats(patch, verbose = TRUE)
}
\arguments{
\item{patch}{patch data}

\item{verbose}{print statistics to screen?}
}
\value{
Dataframe with patch statistics
}
\description{
Helper for reporting patch statistics in build function
}
