% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_stands.R
\name{calculate_adj}
\alias{calculate_adj}
\title{Calculate stands adjacency from input shapefile}
\usage{
calculate_adj(Shapefile, Adjdist = 1, St_id, method = "buffer")
}
\arguments{
\item{Shapefile}{Input shapefile}

\item{Adjdist}{Distance between adjacent stands}

\item{St_id}{Integer vector containing stands IDs}
}
\value{
g igraph
}
\description{
Calculate stands adjacency from input shapefile
}
