% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_stands.R
\name{calculate_dist}
\alias{calculate_dist}
\title{Calculate stands distance from input shapefile}
\usage{
calculate_dist(Shapefile)
}
\arguments{
\item{Shapefile}{Input shapefile}
}
\description{
Calculate stands distance from input shapefile
}
