% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patchmax_r6_aux.R
\name{circleVertices}
\alias{circleVertices}
\title{Generate vertex coordinates for a circle}
\usage{
circleVertices(xc, yc, radius, npoints = 25)
}
\arguments{
\item{xc}{x center}

\item{yc}{y center}

\item{radius}{radius of circle}

\item{npoints}{number of points}
}
\description{
Generate vertex coordinates for a circle
}
