% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patchmax_r6_func.R
\name{create_adj_network}
\alias{create_adj_network}
\title{Calculate stand adjacency and edgewise distances}
\usage{
create_adj_network(geom, id_field, method = "queen", adj_edgelist = NULL)
}
\arguments{
\item{geom}{sf-type geometry with attributes}

\item{id_field}{vector of sf feature ideas (i.e., the stand id)}

\item{method}{adjacency method to use: queen, rook, buffer}

\item{adj_edgelist}{???}
}
\value{
adjacency network saved as an igraph network object
}
\description{
Calculate stand adjacency and edgewise distances
}
