% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patchmax_r6_func.R
\name{dist_func}
\alias{dist_func}
\title{Build graph object used to calculate patch}
\usage{
dist_func(net, objective_field, sdw = 0, epw = 0)
}
\arguments{
\item{net}{igraph object}

\item{objective_field}{name of variable containing objective}

\item{sdw}{numeric between 0 and 1 distance multiplier for traversing low objective stands}

\item{epw}{numeric between 0 and 1 distance multiplier for traversing excluded stands}
}
\value{
cpp graph object
}
\description{
Build graph object used to calculate patch
}
\details{
\code{epw} and \code{sdw} have exponential effects. At 1, distance values are 10x greater than at 0. Setting either value to 0 leaves distances unmodified.
}
