% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forsys_helper.R
\name{forsys_helper_prep_threshold}
\alias{forsys_helper_prep_threshold}
\title{Forsys threshold helper function}
\usage{
forsys_helper_prep_threshold(stand_threshold)
}
\arguments{
\item{stand_threshold}{}
}
\description{
Convert Forsys threshold into format expected by Patchmax
}
