% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patchmax_r6_func.R
\name{net_func}
\alias{net_func}
\title{Calculate stand adjacency and edgewise distances}
\usage{
net_func(geom, id_field, method = "queen")
}
\arguments{
\item{geom}{sf-type geometry with attributes}

\item{id_field}{vector of sf feature ideas (i.e., the stand id)}

\item{method}{adjacency method to use: queen, rook, buffer}
}
\value{
adjacency network saved as an igraph network object
}
\description{
Calculate stand adjacency and edgewise distances
}
