% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patchmax_r6.R
\name{patchmax}
\alias{patchmax}
\title{Patchmax Class}
\description{
Patchmax patch selection object. Use params active binding to
view class parameters or set mutiple parameters at once (using a named list).
}
\examples{
geom <- patchmax::test_forest
pm <- patchmax$new(geom, 'id', 'p1', 'ha', 20000)
pm$params = list(constraint_field = 'p4', constraint_max = 50, area_min=10000)
pm$search(sample_frac = .1, show_progress = T)
pm$build()$record()
pm$plot()

}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{net}}{Get igraph object. Read only}

\item{\code{geom}}{Get sf geometry object. Read only}

\item{\code{best}}{Get pending stand id representing best patch origin. Read only}

\item{\code{pending_stands}}{Get stands in pending patch. Read only}

\item{\code{pending_patch}}{Get pending patch stats. Read only}

\item{\code{patch_stands}}{Get list of recorded stands}

\item{\code{patch_stats}}{Get list of recorded patches}

\item{\code{id_field}}{Get/set stand ID field}

\item{\code{objective_field}}{Get/set objective field}

\item{\code{area_field}}{Get/set area field}

\item{\code{threshold}}{Get/set threshold boolean statement}

\item{\code{exclusion_limit}}{Get/set threshold limit}

\item{\code{threshold_area_adjust}}{Get/set fraction of area to count within excluded stands}

\item{\code{threshold_objective_adjust}}{Get/set fraction of objective to count within excluded stands}

\item{\code{constraint_field}}{Get/set secondary constraint field. Optional}

\item{\code{constraint_max}}{Get/set max value for secondary constraint (e.g., max budget)}

\item{\code{constraint_min}}{Get/set min value for secondary constraint}

\item{\code{area_max}}{Get/set max value for area constraint}

\item{\code{area_min}}{Get/set min value for area constraint}

\item{\code{epw}}{Get/set exclusion penality weight between -1 and 1. Default 0. At 0, patches neither privilege or penalize excluded ares. Values closer to 1 enact a greater cost on projects spanning areas excluded by the project stand threshold. Values less than 0 readily search out excluded areas.}

\item{\code{sdw}}{Get/set spatial distance weight between -1 and 1. Default 0. Patches are highly constrained by distance at -1 and unconstrained by distance at 1.}

\item{\code{params}}{Get/set list of all patch parameters. Argument is a named list if setting multiple parameters at once through params active bindings. Ex: patchmax$params = list(constraint_field = 'constraint1', constraint_max = 1000).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-patchmax-new}{\code{patchmax$new()}}
\item \href{#method-patchmax-build}{\code{patchmax$build()}}
\item \href{#method-patchmax-search}{\code{patchmax$search()}}
\item \href{#method-patchmax-simulate}{\code{patchmax$simulate()}}
\item \href{#method-patchmax-plot}{\code{patchmax$plot()}}
\item \href{#method-patchmax-record}{\code{patchmax$record()}}
\item \href{#method-patchmax-summarize}{\code{patchmax$summarize()}}
\item \href{#method-patchmax-reset}{\code{patchmax$reset()}}
\item \href{#method-patchmax-clone}{\code{patchmax$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-patchmax-new"></a>}}
\if{latex}{\out{\hypertarget{method-patchmax-new}{}}}
\subsection{Method \code{new()}}{
Initialize new patchmax object for building patches
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{patchmax$new(
  geom,
  id_field = NULL,
  objective_field = NULL,
  area_field = NULL,
  area_max = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{geom}}{sf Dataframe like sf object with geomerty}

\item{\code{id_field}}{character Field name containing unique IDs}

\item{\code{objective_field}}{character Field name containing objective values}

\item{\code{area_field}}{character Field name containing area}

\item{\code{area_max}}{numeric Size of patch}

\item{\code{...}}{Additional parameters to pass to \code{patchmax$params}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-patchmax-build"></a>}}
\if{latex}{\out{\hypertarget{method-patchmax-build}{}}}
\subsection{Method \code{build()}}{
Build patch at selected node
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{patchmax$build(node = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{node}}{character. Stand ID used to build patch (optional)}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
If node is NULL, use stand id found during search
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-patchmax-search"></a>}}
\if{latex}{\out{\hypertarget{method-patchmax-search}{}}}
\subsection{Method \code{search()}}{
Search patch origin with highest objective
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{patchmax$search(
  sample_frac = 0.1,
  search_plot = FALSE,
  return_all = FALSE,
  show_progress = FALSE,
  print_errors = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sample_frac}}{numeric Fraction of stands to evaluate (0-1)}

\item{\code{search_plot}}{logical Map search results}

\item{\code{return_all}}{logical Return search results}

\item{\code{show_progress}}{logical Show search progress bar}

\item{\code{print_errors}}{logical Print search errors to console}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-patchmax-simulate"></a>}}
\if{latex}{\out{\hypertarget{method-patchmax-simulate}{}}}
\subsection{Method \code{simulate()}}{
Search, build, and record multiple patches in sequence
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{patchmax$simulate(n_projects = 1, sample_frac = 0.1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n_projects}}{integer. Number of patches to build}

\item{\code{sample_frac}}{numeric. Fraction of stands to evaluate for patches}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-patchmax-plot"></a>}}
\if{latex}{\out{\hypertarget{method-patchmax-plot}{}}}
\subsection{Method \code{plot()}}{
Plot patch and stand map
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{patchmax$plot(
  plot_field = NULL,
  return_plot = FALSE,
  show_origin = FALSE,
  apply_threshold = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{plot_field}}{character Field name to plot}

\item{\code{return_plot}}{logical Return ggplot object}

\item{\code{show_origin}}{logical}

\item{\code{apply_threshold}}{logical}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-patchmax-record"></a>}}
\if{latex}{\out{\hypertarget{method-patchmax-record}{}}}
\subsection{Method \code{record()}}{
Record selected patch
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{patchmax$record(patch_id = NULL, enforce_constraint = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{patch_id}}{integer/character Patch name. If null, add one to highest}

\item{\code{enforce_constraint}}{logical Apply secondary constraint}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-patchmax-summarize"></a>}}
\if{latex}{\out{\hypertarget{method-patchmax-summarize}{}}}
\subsection{Method \code{summarize()}}{
Summarize recorded patches
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{patchmax$summarize(group_vars = NULL, sum_vars = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{group_vars}}{character vector Field names to group by}

\item{\code{sum_vars}}{characcter vector Field naems to summarize}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-patchmax-reset"></a>}}
\if{latex}{\out{\hypertarget{method-patchmax-reset}{}}}
\subsection{Method \code{reset()}}{
Reset recorded patches
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{patchmax$reset(patch_id = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{patch_id}}{optional. Patch ID to delete}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
If blank, delete all patches. If negative, delete that number of the most recent patches. Else, delete patch ID equal argument.
Update edgelist distances
Update network adjacency network object
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-patchmax-clone"></a>}}
\if{latex}{\out{\hypertarget{method-patchmax-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{patchmax$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
