% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patchmax_r6.R
\name{patchmax_generator}
\alias{patchmax_generator}
\title{Patchmax Class}
\description{
Patchmax patch selection object. Use params active binding to
view class parameters or set mutiple parameters at once (using a named list).
}
\examples{
geom <- patchmax::test_forest
pm <- patchmax$new(geom, 'stand_id', 'priority1', 'area_ha', 10000)
pm <- pm$params = list(constraint_field = 'priority4', constraint_max = 6, area_min=800)
pm$search(sample_frac = 1, plot_search = T)
pm$build()$plot(show_seed = TRUE)$record()

}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{net}}{Get igraph object. Read only}

\item{\code{geom}}{Get sf geometry object. Read only}

\item{\code{best}}{Get pending stand id representing best patch origin. Read only}

\item{\code{patch}}{Get stands in pending patch. Read only}

\item{\code{id_field}}{Get/set stand ID field}

\item{\code{objective_field}}{Get/set objective field}

\item{\code{area_field}}{Get/set area field}

\item{\code{threshold}}{Get/set threshold boolean statement}

\item{\code{threshold_area_adjust}}{Get/set fraction of area to count within excluded stands}

\item{\code{threshold_objective_adjust}}{Get/set fraction of objective to count within excluded stands}

\item{\code{constraint_field}}{Get/set secondary constraint field. Optional}

\item{\code{constraint_max}}{Get/set max value for secondary constraint (e.g., max budget)}

\item{\code{constraint_min}}{Get/set min value for secondary constraint}

\item{\code{area_max}}{Get/set max value for area constraint}

\item{\code{area_min}}{Get/set min value for area constraint}

\item{\code{epw}}{Get/set exclusion penality weight between -1 and 1. Default 0. At 0, patches neither privilege or penalize excluded ares. Values closer to 1 enact a greater cost on projects spanning areas excluded by the project stand threshold. Values less than 0 readily search out excluded areas.}

\item{\code{sdw}}{Get/set spatial distance weight between -1 and 1. Default 0. Patches are highly constrained by distance at -1 and unconstrained by distance at 1.}

\item{\code{params}}{Get/set list of all patch parameters. Arguement is a named list if setting mutiple parameters at once through params active bindings. Ex: patchmax$params = list(constraint_field = 'constraint1', constraint_max = 1000).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-patch_generator-new}{\code{patchmax_generator$new()}}
\item \href{#method-patch_generator-build}{\code{patchmax_generator$build()}}
\item \href{#method-patch_generator-search}{\code{patchmax_generator$search()}}
\item \href{#method-patch_generator-simulate}{\code{patchmax_generator$simulate()}}
\item \href{#method-patch_generator-plot}{\code{patchmax_generator$plot()}}
\item \href{#method-patch_generator-record}{\code{patchmax_generator$record()}}
\item \href{#method-patch_generator-describe}{\code{patchmax_generator$describe()}}
\item \href{#method-patch_generator-reset}{\code{patchmax_generator$reset()}}
\item \href{#method-patch_generator-clone}{\code{patchmax_generator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-patch_generator-new"></a>}}
\if{latex}{\out{\hypertarget{method-patch_generator-new}{}}}
\subsection{Method \code{new()}}{
Initialize new patchmax object for building patches
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{patchmax_generator$new(
  geom,
  id_field = NULL,
  objective_field = NULL,
  area_field = NULL,
  area_max = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{geom}}{sf Dataframe like sf object with geomerty}

\item{\code{id_field}}{character Field name containing unique IDs}

\item{\code{objective_field}}{character Field name containing objective values}

\item{\code{area_field}}{character Field name containing area}

\item{\code{area_max}}{numeric Size of patch}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-patch_generator-build"></a>}}
\if{latex}{\out{\hypertarget{method-patch_generator-build}{}}}
\subsection{Method \code{build()}}{
Build patch at selected node
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{patchmax_generator$build(node = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{node}}{character. Stand ID used to build patch (optional)}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
If node is NULL, use stand id found during search
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-patch_generator-search"></a>}}
\if{latex}{\out{\hypertarget{method-patch_generator-search}{}}}
\subsection{Method \code{search()}}{
Search patch origin with highest objective
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{patchmax_generator$search(
  sample_frac = 0.1,
  return_all = FALSE,
  show_progress = FALSE,
  plot_search = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sample_frac}}{numeric Fraction of stands to evaluate (0-1)}

\item{\code{return_all}}{logical Return search results}

\item{\code{show_progress}}{logical Show search progress bar}

\item{\code{plot_search}}{logical Map search results}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-patch_generator-simulate"></a>}}
\if{latex}{\out{\hypertarget{method-patch_generator-simulate}{}}}
\subsection{Method \code{simulate()}}{
Search, build, and record multiple patches in sequence
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{patchmax_generator$simulate(n_projects = 1, sample_frac = 0.1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n_projects}}{integer. Number of patches to build}

\item{\code{sample_frac}}{numeric. Fraction of stands to evaluate for patches}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-patch_generator-plot"></a>}}
\if{latex}{\out{\hypertarget{method-patch_generator-plot}{}}}
\subsection{Method \code{plot()}}{
Plot patch and stand map
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{patchmax_generator$plot(
  plot_field = NULL,
  return_plot = FALSE,
  enforce_constraint = TRUE,
  show_seed = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{plot_field}}{character Field name to plot}

\item{\code{return_plot}}{logical Return ggplot object}

\item{\code{enforce_constraint}}{logical Apply secondary constraint}

\item{\code{show_seed}}{logical Show origin used to build selected patch?}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-patch_generator-record"></a>}}
\if{latex}{\out{\hypertarget{method-patch_generator-record}{}}}
\subsection{Method \code{record()}}{
Record selected patch
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{patchmax_generator$record(patch_id = NULL, enforce_constraint = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{patch_id}}{integer/character Patch name. If null, add one to highest}

\item{\code{enforce_constraint}}{logical Apply secondary constraint}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-patch_generator-describe"></a>}}
\if{latex}{\out{\hypertarget{method-patch_generator-describe}{}}}
\subsection{Method \code{describe()}}{
Summarize recorded patches
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{patchmax_generator$describe()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-patch_generator-reset"></a>}}
\if{latex}{\out{\hypertarget{method-patch_generator-reset}{}}}
\subsection{Method \code{reset()}}{
Reset recorded patches
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{patchmax_generator$reset()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-patch_generator-clone"></a>}}
\if{latex}{\out{\hypertarget{method-patch_generator-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{patchmax_generator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
