% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_stands.R
\name{read_adj}
\alias{read_adj}
\title{Read stands adjacency from input matrix}
\usage{
read_adj(Adjlist)
}
\arguments{
\item{Adjlist}{Adjacency list}
}
\value{
g igraph
}
\description{
Read stands adjacency from input matrix
}
