% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patchmax_r6_func.R
\name{sample_frac}
\alias{sample_frac}
\title{Samples stands on regular spatial grid}
\usage{
sample_frac(geom, sample_frac, spatial_grid = TRUE)
}
\arguments{
\item{geom}{sf. Stand geometry}

\item{sample_frac}{numeric. Fraction of stands to evaluate}

\item{spatial_grid}{logical. Sample at regular spatial intervals?}
}
\description{
Samples stands on regular spatial grid
}
