% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patchmax_r6_func.R
\name{sample_func}
\alias{sample_func}
\title{Samples stands on regular spatial grid}
\usage{
sample_func(geom, sample_frac, id_field, spatial_grid = TRUE, rng_seed = NULL)
}
\arguments{
\item{geom}{sf. Stand geometry}

\item{sample_frac}{numeric. Fraction of stands to evaluate}

\item{id_field}{???}

\item{spatial_grid}{logical. Sample at regular spatial intervals?}

\item{rng_seed}{???}
}
\description{
Samples stands on regular spatial grid
}
