% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patchmax_r6_func.R
\name{search_best_func}
\alias{search_best_func}
\title{Evaluate objective score for all or fraction of patch stand seeds}
\usage{
search_best_func(
  net,
  cpp_graph,
  objective_field,
  a_max,
  a_min,
  c_max = Inf,
  c_min = -Inf,
  sample_frac = 1,
  return_all = FALSE,
  show_progress = FALSE,
  sample_type = "spatial"
)
}
\arguments{
\item{net}{igraph graph object}

\item{cpp_graph}{cpp graph object}

\item{objective_field}{name of field containing objective values}

\item{a_max}{patch size}

\item{sample_frac}{fraction of stands to evaluate}
}
\description{
Evaluate objective score for all or fraction of patch stand seeds
}
\details{
Calculates potential patches for all or fraction of landscape stands
in order to identify the initial seed that leads to the highest total
objective score.
}
