% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patchmax_r6_func.R
\name{search_func}
\alias{search_func}
\title{Evaluate objective score for all or fraction of patch stand seeds}
\usage{
search_func(
  net,
  cpp_graph,
  nodes = NULL,
  objective_field,
  a_max,
  a_min,
  c_max = Inf,
  c_min = -Inf,
  t_limit = 0,
  return_all = FALSE,
  show_progress = FALSE,
  print_errors = FALSE
)
}
\arguments{
\item{net}{igraph graph object}

\item{cpp_graph}{cpp graph object}

\item{nodes}{}

\item{objective_field}{name of field containing objective values}

\item{a_max}{patch size}

\item{a_min}{}

\item{c_max}{}

\item{c_min}{}

\item{t_limit}{}

\item{return_all}{}

\item{show_progress}{}

\item{print_errors}{}
}
\description{
Evaluate objective score for all or fraction of patch stand seeds
}
\details{
Calculates potential patches for all or fraction of landscape stands in order to identify the initial seed that leads to the highest total objective score.
}
