% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_projects.R
\name{simulate_projects}
\alias{simulate_projects}
\title{Simulate landscape projects}
\usage{
simulate_projects(
  St_id,
  St_adj,
  St_area,
  St_objective,
  P_size,
  P_size_slack,
  P_number,
  St_threshold = NULL,
  St_threshold_value = NULL,
  P_constraint = NULL,
  P_constraint_max_value = NULL,
  P_constraint_min_value = NULL,
  Candidate_min_size = NULL
)
}
\arguments{
\item{St_id}{Numeric vector of integer stands IDs.}

\item{St_adj}{Adjacency input graph. From calculate_adj or read_adj functions}

\item{St_area}{Numeric vector of stands area}

\item{St_objective}{Numeric vector of stands objective}

\item{P_size}{Project size}

\item{P_size_slack}{Project size slack, between 0 and 1, Where 0 means no deviation.}

\item{P_number}{Number of projects to simulate}

\item{St_threshold}{Numeric vector of stands threshold value.Coupled with St_threshold_value. If NULL, then stand threshold is not applied.}

\item{St_threshold_value}{Stands threshold lower value.}

\item{P_constraint}{Numeric vector of stands value for the project constraint.Coupled with P_constraint_max_value and P_constraint_min_value. If NULL, then project constraint is not applied.}

\item{P_constraint_max_value}{Project constraint upper value.}

\item{P_constraint_min_value}{Project constraint lower value.}

\item{Candidate_min_size}{Minimal size for project coded with types 2 and 3. If NULL, then the value = ‘0.25*P_size’ is used by default. Project type codes interpret if the project is valid (i.e. if P_size apply), constrained (i.e. if P_constraint apply) or none of them. Where type 0 are valid non-constrained projects (i.e. only P_size apply), type 1 are valid constrained projects (i.e. P_size and P_constraint apply), type 2 are invalid constrained projects (i.e. P_size does not apply but P_constraint apply) and type 3 are non-optimized projects that can be invalid and/or non-constrained (i.e. P_size and/or P_constraint do not apply).}
}
\value{

}
\description{
Simulate landscape projects
}
