% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_projects.R
\name{simulate_projects_func}
\alias{simulate_projects_func}
\title{Internal simulate landscape projects function}
\usage{
simulate_projects_func(
  St_id,
  St_adj,
  St_area,
  St_objective,
  St_seed,
  P_size,
  P_size_slack,
  P_size_ceiling,
  P_number,
  St_threshold,
  St_threshold_value,
  St_distances,
  SDW,
  P_constraint,
  P_constraint_max_value,
  P_constraint_min_value,
  Candidate_min_size
)
}
\description{
Internal simulate landscape projects function
}
\keyword{internal}
