% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patchmax_r6_func.R
\name{st_queen}
\alias{st_queen}
\title{Helper function to calculate queen adjacency in geometry}
\usage{
st_queen(geom)
}
\arguments{
\item{geom}{sf-class geometry}
}
\description{
Helper function to calculate queen adjacency in geometry
}
