% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patchmax_r6_func.R
\name{st_rook}
\alias{st_rook}
\title{Helper function to calculate rook adjacency in geometry}
\usage{
st_rook(geom)
}
\arguments{
\item{geom}{sf-class geometry}
}
\description{
Helper function to calculate rook adjacency in geometry
}
