% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patchmax_r6_func.R
\name{threshold_func}
\alias{threshold_func}
\title{Set threshold for patch inclusion}
\usage{
threshold_func(net, include, area_adjust = 0, objective_adjust = 0)
}
\arguments{
\item{net}{igraph adjacency network}

\item{include}{logical vector of same length as the count of network nodes}

\item{area_adjust}{value 0-1 represent area 'cost' of excluded stands}

\item{objective_adjust}{value 0-1, percent contribution of excluded stands
to patch score}
}
\value{
igraph adjacency network
}
\description{
Set threshold for patch inclusion
}
\details{
By default, excluded stands don't count towards the total project
objective or the total project size. Setting the area and objective adjust
to a fraction greater than 0 changes this behavior, allowing for a fraction
of the both values for excluded stands to be counted.
}
