% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patchmax_r6_func.R
\name{build_func}
\alias{build_func}
\title{Build patch of specified size while minimizing modified distance costs}
\usage{
build_func(
  seed,
  edge_dat,
  node_dat,
  a_max,
  a_min = -Inf,
  c_max = Inf,
  c_min = -Inf,
  c_enforce = TRUE
)
}
\arguments{
\item{seed}{character. Node id to start building patch}

\item{edge_dat}{graph object built with cppRouting}

\item{node_dat}{data.table Stand data table to evaluate}

\item{a_max}{numeric. Maximum size of patch}

\item{a_min}{numeric. Minimum size of patch}

\item{c_max}{numeric. Maximum secondary constraint of patch}

\item{c_min}{numeric. Minimum secondary constraint of patch}

\item{c_enforce}{logical. Should patches outside of constraint be excluded?}
}
\value{
data frame of nearest nodes arranged by distance
}
\description{
Build patch of specified size while minimizing modified distance costs
}
