% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patchmax_r6_func.R
\name{build_graph_func}
\alias{build_graph_func}
\title{Build graph object used to calculate patch}
\usage{
build_graph_func(net, objective_field, sdw = 0, epw = 0)
}
\arguments{
\item{net}{igraph object}

\item{objective_field}{name of variable containing objective}

\item{sdw}{numeric between -1 and 1 controlling flexibility}

\item{epw}{numeric between -1 and 1 distance multiplier for traversing excluded stands}
}
\value{
cpp graph object
}
\description{
Build graph object used to calculate patch
}
\details{
\code{epw} values less than 1 will preferentially select excluded stands
while values greater than 1 avoids excluded stands.
}
