% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patchmax_r6_func.R
\name{search_func}
\alias{search_func}
\title{Evaluate objective score for all or fraction of patch stand seeds}
\usage{
search_func(
  net,
  edge_dat,
  node_dat,
  nodes = NULL,
  objective_field,
  a_max,
  a_min,
  c_max = Inf,
  c_min = -Inf,
  t_limit = 0,
  show_progress = FALSE,
  print_errors = FALSE
)
}
\arguments{
\item{net}{igraph graph object}

\item{nodes}{which nodes to evaluate as potential patches}

\item{objective_field}{name of field containing objective values}

\item{a_max}{maxmimum patch size}

\item{a_min}{minimum patch size (only applicable if constraint is present)}

\item{c_max}{maximum constraint value}

\item{c_min}{minimum constraint value}

\item{t_limit}{maximum portion (0-1) of patch area excluded by threshold}

\item{show_progress}{logical Show progress bar}

\item{print_errors}{logical Print reason for invalid search (for debugging)}

\item{cpp_graph}{cpp graph object}

\item{return_all}{logical Return search values for all nodes evaluated}
}
\description{
Evaluate objective score for all or fraction of patch stand seeds
}
\details{
Calculates potential patches for all or fraction of landscape stands in order to identify the initial seed that leads to the highest total objective score.
}
