% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patchmax_r6_wrapper.R
\name{simulate_projects}
\alias{simulate_projects}
\title{Simulate landscape projects (wrapper)}
\usage{
simulate_projects(
  geom,
  St_id,
  St_area,
  St_objective,
  St_seed = NULL,
  P_size,
  P_size_min = -Inf,
  P_number = 1,
  P_ceiling = NULL,
  P_ceiling_max = Inf,
  St_threshold = NULL,
  SDW = 0.5,
  EPW = 0.5,
  exclusion_limit = 0.5,
  P_constraint = NULL,
  P_constraint_max_value = Inf,
  P_constraint_min_value = -Inf,
  sample_frac = 0.1,
  sample_seed = NULL
)
}
\arguments{
\item{geom}{sf. Geometry representing stands}

\item{St_id}{vector. stands IDs. Converted to character}

\item{St_area}{numeric vector. Stands area}

\item{St_objective}{numeric vector. Stands objective values}

\item{St_seed}{numeric vector. Stands IDs seeds to search (n)}

\item{P_size}{numeric. Project size}

\item{P_size_min}{numeric. Minimum project size (when \code{P_constraint} is specified).}

\item{P_number}{integer. Number of patches to simulate}

\item{St_threshold}{character. Boolean statement on stand availability (e.g., 'field > 0.5')}

\item{SDW}{numeric 0-1. Objective weight parameter. Higher objectives are preferentially sought at higher SDW. Default is 0.5.}

\item{EPW}{numeric 0-1. Exclusion weight parameter. Excluded stands are avoided at higher EPW. Default is 0.5.}

\item{exclusion_limit}{numeric 0-1. Maximum percent of patch area that can be excluded.}

\item{P_constraint}{numeric vector. Stands values for the project constraint. If NULL, then  no secondary project constraint is not applied.}

\item{P_constraint_max_value}{numeric. Project constraint upper value. Default is Inf}

\item{P_constraint_min_value}{numeric. Project constraint lower value. Default is -Inf}

\item{sample_frac}{numeric 0-1. Portion of stands to search}
}
\value{
list with first element describing patch-level stats and the second
element describing stand-level stats.
}
\description{
Simulate landscape projects (wrapper)
}
\details{
This function is meant as a API for calling patchmax within ForSys
(or another environment) using a minimal set of parameters.
}
